Pass reductions through shuffles
This is here mostly for comments.  In many (but not all) cases we should
be able to pass reductions through shuffle operations.  For example:

```python
len(df.set_index(...)) == len(df)
```

This is easy to write, although we probably want to allow-list these
rather than disallow them as we do today (IdxMax is a good example of a
reduction where this isn't doable).

Additionally, we should also be able to do this with a projection in the
middle:

```python
df.set_index(...).x.sum() == df.x.sum()
```

However we can't actually express this with today's optimizations
because `_simplify_up` doesn't reach far enough up.  Interestingly this
is a case where matchpy would have given us more than we can achieve
with our current optimization infrastructure.

Probably short-term we should do the simple things and merge a cleaner
version of this in.  I thought I'd leave it in this state though for
discussion for a moment first.